requirements.txt
attrs==21.2.0
cached-property==1.5.2
certifi==2021.5.30
charset-normalizer==2.0.4
et-xmlfile==1.1.0
idna==3.2
isodate==0.6.0
lxml==4.6.3
numpy==1.21.2
openpyxl==3.0.7
pandas==1.3.2
platformdirs==2.3.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
requests-file==1.5.1
requests-toolbelt==0.9.1
schedule==1.1.0
six==1.16.0
urllib3==1.26.6
zeep==4.1.0
